requirements.txt
flask
ultralytics
Pillow
gunicorn
